% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MDE.R
\name{calculate_MDE}
\alias{calculate_MDE}
\title{Calculate minimal detectable effect for binary variable}
\usage{
calculate_MDE(number_of_events, alpha = 0.05, v = 0.5)
}
\arguments{
\item{number_of_events}{Number of events}

\item{alpha}{1-sided alpha (Type I error)}

\item{v}{Proportion of treatment allocation (default is 1:1)}
}
\value{
Minimal detectable effect
}
\description{
Calculate the minimal detectable effect (Hazard Ratio) given an observed number of events for a binary variable (based on Schoenfeld's formula)
}
