% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MDE_continuous_variable.R
\name{calculate_MDE_biomarker}
\alias{calculate_MDE_biomarker}
\title{Calculate minimal detectable effect for a continuous variable}
\usage{
calculate_MDE_biomarker(number_of_events, alpha = 0.05, sd = 1)
}
\arguments{
\item{number_of_events}{Number of events}

\item{alpha}{1-sided alpha (Type I error)}

\item{sd}{Standarad deviation of the continuous covariate (default is 1 assuming the continuous marker is standardized)}
}
\value{
Minimal detectable effect
}
\description{
Calculate the minimal detectable effect (Hazard Ratio) given an observed number of events for a continuous variable (based on Schoenfeld's formula)
Reference: "Sample-Size Calculations for the Cox Proportional Hazards Regression Model with Nonbinary Covariates" by F.Y. Hsieh, PhD, and Philip W. Lavori, PhD (2000)
}
