% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MDE_WinRatio.R
\name{calculate_MDE_WinRatio}
\alias{calculate_MDE_WinRatio}
\title{Calculate minimal detectable effect for binary variable using the WinRatio}
\usage{
calculate_MDE_WinRatio(sample_size, P_tie, alpha = 0.05, v = 0.5)
}
\arguments{
\item{sample_size}{Total Sample Size}

\item{P_tie}{Proportion of ties}

\item{alpha}{1-sided alpha (Type I error)}

\item{v}{Proportion of treatment allocation (default is 1:1)}
}
\value{
Minimal detectable effect
}
\description{
Calculate the minimal detectable effect (Win Ratio) given the total sample size (Yu and Ganju, Statistics in Medicine 2021)
}
