% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_power.R
\name{calculate_power}
\alias{calculate_power}
\title{Calculate Power}
\usage{
calculate_power(number_of_events, HR, alpha, v)
}
\arguments{
\item{number_of_events}{Number of Events}

\item{HR}{Assumed Effect Size}

\item{alpha}{1-sided Alpha (Type I error)}

\item{v}{Proportion}
}
\value{
Power
}
\description{
Based on Schoenfeld's formula
}
