% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_rel.R
\name{delta_rel}
\alias{delta_rel}
\title{Interaction effect relative to the effect size for the total population}
\usage{
delta_rel(HR_s, HR_c, HR_overall)
}
\arguments{
\item{HR_s}{HR in the Subgroup of Interest}

\item{HR_c}{HR in the Comparative Subgroup}

\item{HR_overall}{HR of the Overall Cohort}
}
\value{
Interaction Effect Relative to the Effect Size for the Total Population
}
\description{
Based on Alosh, Huque, Koch  (2015)
}
