% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_total_HR.R
\name{find_total_HR}
\alias{find_total_HR}
\title{Calculate the HR of the Total Cohort given HRs, Number of Pts Included, and Target Events}
\usage{
find_total_HR(
  n_CVD,
  n_MRF,
  HR_CVD,
  HR_MRF,
  target_events,
  ratio_event_rate_CVDvsMRF
)
}
\arguments{
\item{n_CVD}{Number of patients with cardiovascular disease}

\item{n_MRF}{Number of patients with multiple risk factors}

\item{target_events}{Number of targeted events}

\item{ratio_event_rate_CVDvsMRF}{Ratio btw. event rate of CVD versus event rate of MRF}
}
\value{
Total hazard ratio
}
\description{
Based on Alosh, Huque, Koch  (2015)
}
