% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_confint.R, R/format_pval.R
\name{format_pval}
\alias{format_pval}
\title{Format p-values}
\usage{
format_pval(x)

format_pval(x)
}
\arguments{
\item{x}{Numeric variable}

\item{HR}{Point estimate}

\item{HR_lower}{Lower CI}

\item{HR_upper}{Upper CI}
}
\value{
Formatted confidence interval as a string

P-value as string
}
\description{
Format CI with 95%-CI

Format p-values: p-values <= 0.10 are printed with 3 digits, otherwise 2 digits
}
