% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ARR_confint.R
\name{get_ARR_confint}
\alias{get_ARR_confint}
\title{Calculate Absolute Risk Difference, the SE, and confidence interval}
\usage{
get_ARR_confint(trt_n, trt_N, placebo_n, placebo_N)
}
\arguments{
\item{trt_n}{Number of events with Treatment}

\item{trt_N}{Total number of patients in the treatment group}

\item{placebo_n}{Number of events with Placebo}

\item{placebo_N}{Total number of patients in the placebo group}
}
\value{
ARR with CI
}
\description{
Calculate Absolute Risk Difference, the SE, and confidence interval
}
