% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_events_Cox.R
\name{number_events_Cox}
\alias{number_events_Cox}
\title{Calculate number of events}
\usage{
number_events_Cox(alpha, beta, HR, v = 0.5)
}
\arguments{
\item{alpha}{1-sided alpha (Type I error)}

\item{beta}{Type II error}

\item{HR}{Assumed effect size}

\item{v}{Proportion of patients that are assigned to the intervention arm  (Default is  0.5)}
}
\value{
Number of events required
}
\description{
Based on Schoenfeld's formula
}
