% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_Cox.R
\name{power_Cox}
\alias{power_Cox}
\title{Calculate Power}
\usage{
power_Cox(number_of_events, HR, alpha, v)
}
\arguments{
\item{number_of_events}{Number of Events}

\item{HR}{Assumed Effect Size}

\item{alpha}{1-sided Alpha (Type I error)}

\item{v}{Proportion}
}
\value{
Power
}
\description{
Based on Schoenfeld's formula
}
