% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_nb.R
\name{power_nb}
\alias{power_nb}
\title{Calculate sample size or power for comparing 2 negative binomial rates
(either n or power needs to included)}
\usage{
power_nb(
  n = NULL,
  r0,
  r1,
  RR,
  k,
  duration = 1,
  ssize.ratio = 1,
  alpha = 0.025,
  power = NULL,
  approach = 3
)
}
\arguments{
\item{n}{Sample Size; Leave NULL to calculate power}

\item{r0}{Event Rate (Placebo)}

\item{r1}{Event Rate (Treatment); Alernatively, enter RR}

\item{RR}{Rate Ratio}

\item{k}{Negative binomial dispersion parameter (see dispersion_nb Function)}

\item{duration}{Average treatment duration}

\item{ssize.ratio}{Sample allocation ratio n1/n0 (Default is  1)}

\item{alpha}{1-sided alpha (Type I error)}

\item{approach}{1 (Use reference group rate), 2 (Use true rates), or 3 (Use ML estimation); see Zhu and Lakkis (2014).}

\item{beta}{Type II error; Leave NULL to calculate sample size}
}
\value{
Either power of Number of events reported
}
\description{
Calculate sample size or power for comparing 2 negative binomial rates
(either n or power needs to included)
}
