% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_subgroup.R
\name{power_subgroup}
\alias{power_subgroup}
\title{Calculate power for subgroup interaction}
\usage{
power_subgroup(alpha_s, delta_rel, alpha, beta, K)
}
\arguments{
\item{alpha_s}{1-sided alpha (Type I error) for subgroup interaction}

\item{delta_rel}{Interaction effect relative to the overall effect [(log(HR_s - log(delta_c)) / log(HR_overall)], see separate function}

\item{alpha}{Type I error of the study (one-sided alpha)}

\item{beta}{Type II error of the study}

\item{K}{Proportion}
}
\value{
Power of subgroup interaction
}
\description{
Based on Alosh, Huque, Koch  (2015)
}
