% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_numbers_of_events.R
\name{required_events}
\alias{required_events}
\title{Calculate number of events}
\usage{
required_events(alpha, beta, HR, v = 0.5)
}
\arguments{
\item{alpha}{1-sided alpha (Type I error)}

\item{beta}{Type II error}

\item{HR}{Assumed effect size}

\item{v}{Proportion of patients that are assigned to the intervention arm  (Default is  0.5)}
}
\value{
Number of events required
}
\description{
Based on Schoenfeld's formula
}
