% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_numbers_of_events_continuous_variable.R
\name{required_events_biomarker}
\alias{required_events_biomarker}
\title{Calculate number of events}
\usage{
required_events_biomarker(alpha, beta, HR, sigma)
}
\arguments{
\item{alpha}{1-sided Alpha (Type I error)}

\item{beta}{Type II error}

\item{HR}{Hazard ratio associated with a one-unit change in a continuous covariate (X1)}

\item{sigma}{standard deviation of the continuous covariate (X1)}
}
\value{
Number of events required
}
\description{
#' Calculate the number of events for a continuous covariate (based on Schoenfeld's formula)
Reference: "Sample-Size Calculations for the Cox Proportional Hazards Regression Model with Nonbinary Covariates" by F.Y. Hsieh, PhD, and Philip W. Lavori, PhD (2000)
}
