% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_winratio.R
\name{sample_size_winratio}
\alias{sample_size_winratio}
\title{Calculate number of events using the winratio. Sample size calculatin based on Yu and Ganju, Statistics in Medicine 2021}
\usage{
sample_size_winratio(alpha, beta, WR_true, P_tie, v = 0.5)
}
\arguments{
\item{alpha}{1-sided alpha (Type I error)}

\item{beta}{Type II error}

\item{WR_true}{Assumed effect size}

\item{P_tie}{Proportion of ties}

\item{v}{Proportion of patients that are assigned to the intervention arm  (Default is  0.5)}
}
\value{
Total sample size required
}
\description{
Based on Schoenfeld's formula
}
