% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_nb.R
\name{dispersion_nb}
\alias{dispersion_nb}
\title{Estimate the negative binomial dispersion parameter k (according to Zhu H and Lakkis H, Stat Med 2014)}
\usage{
dispersion_nb(r, CI_lower, CI_upper, exposure_time)
}
\arguments{
\item{r}{Event rate per patient year}

\item{CI_lower}{Lower confidence interval of r}

\item{exposure_time}{Total exposure time in years}

\item{CI_lower}{Upper confidence interval of r}
}
\value{
Negative binomial dispersion parameter k
}
\description{
Estimate the negative binomial dispersion parameter k (according to Zhu H and Lakkis H, Stat Med 2014)
}
