% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{Rounding of Numbers}
\usage{
round2(x, digits = 1)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{integer indicating the number of decimal places}
}
\description{
R  “rounds to even” when rounding off a 5. This formula rounds numbers ending in .5  up ("rounding in commerce").
This formula is based on http://alandgraf.blogspot.com/2012/06/rounding-in-r.html
}
